+++ packages/terra-action-footer/tests/wdio/theme/centered-action-footer-theme-spec.js
@@ -7,12 +7,18 @@ describe('Centered action footer theme', () => {
   describe('Empty', () => {
     beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/empty-centered-action-footer'));
 
-    Terra.should.themeEachCustomProperty(withoutActionsThemeables);
+    Terra.should.themeCombinationOfCustomProperties({
+      testName: 'themed',
+      properties: withoutActionsThemeables,
+    });
   });
 
   describe('Multiple Actions', () => {
     beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/multiple-action-centered-action-footer'));
 
-    Terra.should.themeEachCustomProperty(withActionsThemeables);
+    Terra.should.themeCombinationOfCustomProperties({
+      testName: 'themed',
+      properties: withActionsThemeables,
+    });
   });
 });