+++ packages/terra-button/src/terra-dev-site/doc/example/ButtonDisabled.jsx
@@ -9,6 +9,7 @@ const ButtonDisabled = () => (
   <div>
     <Button text="Disabled Button" isDisabled style={buttonStyle} />
     <Button text="Disabled Emphasis" variant="emphasis" isDisabled style={buttonStyle} />
+    <Button text="Disabled Ghost" variant="ghost" isDisabled style={buttonStyle} />
     <Button text="Disabled De-emphasis" variant="de-emphasis" isDisabled style={buttonStyle} />
     <Button text="Disabled Action" variant="action" isDisabled icon={<IconEdit />} style={buttonStyle} />
     <Button text="Disabled Utility" variant="utility" isDisabled icon={<IconEdit />} style={buttonStyle} />