+++ packages/terra-form-input/tests/wdio/form-input-spec.js
@@ -65,8 +65,11 @@ describe('Form-Input', () => {
 
       Terra.should.beAccessible({ viewports });
       Terra.should.matchScreenshot({ viewports });
-      Terra.should.themeEachCustomProperty({
-        '--terra-form-input-hover-border': '20rem dashed blue',
+      Terra.should.themeCombinationOfCustomProperties({
+        testName: 'themed',
+        properties: {
+          '--terra-form-input-hover-border': '20rem dashed blue',
+        },
       });
     });
 
@@ -93,8 +96,11 @@ describe('Form-Input', () => {
 
       Terra.should.beAccessible({ viewports });
       Terra.should.matchScreenshot({ viewports });
-      Terra.should.themeEachCustomProperty({
-        '--terra-form-input-color': 'red',
+      Terra.should.themeCombinationOfCustomProperties({
+        testName: 'themed',
+        properties: {
+          '--terra-form-input-color': 'red',
+        },
       });
     });
 
@@ -177,8 +183,11 @@ describe('Form-Input', () => {
 
       Terra.should.beAccessible({ viewports });
       Terra.should.matchScreenshot({ viewports });
-      Terra.should.themeEachCustomProperty({
-        '--terra-form-input-hover-border': '20rem dashed red',
+      Terra.should.themeCombinationOfCustomProperties({
+        testName: 'themed',
+        properties: {
+          '--terra-form-input-hover-border': '20rem dashed red',
+        },
       });
     });