+++ packages/terra-scroll/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.8.0 - (May 1, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.7.0 - (April 16, 2019)
 ------------------
 ### Changed